#9 use facade object for worldstate and node service
the $resource objects depend on the current settings of the IcmmSettings
service and need to be recreated every time those settings changes. to
get notified about changes, the Worlstate and Nodes service register a
a callback function to the IcmmSettings service. In fact the $resource
objects need to be recreated, the service factories return a facade
object